# admin-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "i18n-flow 是一个用于管理多语言翻译的系统。",
        "title": "i18n-flow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/languages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统中所有可用的语言列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "获取所有语言",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.LanguageResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "在系统中添加一种新的语言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "创建新语言",
                "parameters": [
                    {
                        "description": "语言信息",
                        "name": "language",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LanguageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service.LanguageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/languages/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新系统中现有语言的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "更新语言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "语言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "语言信息",
                        "name": "language",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LanguageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.LanguageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从系统中删除指定的语言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "语言管理"
                ],
                "summary": "删除语言",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "语言ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.LanguageRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "语言代码",
                    "type": "string",
                    "example": "fr"
                },
                "is_default": {
                    "description": "是否为默认语言",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "语言名称",
                    "type": "string",
                    "example": "Français"
                }
            }
        },
        "service.LanguageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "语言代码",
                    "type": "string",
                    "example": "fr"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-04-01 12:00:00"
                },
                "id": {
                    "description": "语言ID",
                    "type": "integer",
                    "example": 1
                },
                "is_default": {
                    "description": "是否为默认语言",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "语言名称",
                    "type": "string",
                    "example": "Français"
                },
                "status": {
                    "description": "语言状态",
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-04-01 12:00:00"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}